Video_Dynamics_Prediction/requirements.txt
Package           Version
----------------- ---------
absl-py           0.12.0
cycler            0.10.0
dm-env            1.4
dm-tree           0.1.6
enum34            1.1.10
imageio           2.9.0
joblib            1.0.1
kiwisolver        1.3.1
matplotlib        3.4.1
mock              4.0.3
numpy             1.20.2
opencv-python     4.5.1.48
pandas            1.2.4
Pillow            8.2.0
pip               20.2.3
pyparsing         2.4.7
python-dateutil   2.8.1
pytz              2021.1
scikit-learn      0.24.1
scipy             1.6.2
setproctitle      1.2.2
setuptools        49.2.1
six               1.15.0
spriteworld       1.0.2
threadpoolctl     2.1.0
torch             1.4.0+cpu
torchaudio        0.8.1
torchvision       0.5.0+cpu
tqdm              4.60.0
typing-extensions 3.7.4.3
